+++ incubator/lnurl-java/lnurl-core/src/main/java/org/tbk/lnurl/auth/LnurlAuth.java
@@ -22,7 +22,7 @@ public interface LnurlAuth {
 
     enum Action {
         REGISTER("register"), // service will create a new account linked to user's linkingKey.
-        LOGIN("login"), // service will login user to an existing account linked to user's linkingKey.
+        LOGIN("login"), // service will log in user to an existing account linked to user's linkingKey.
         LINK("link"), // service will link a user provided linkingKey to user's existing account (if account was not originally created using lnurl-auth).
         AUTH("auth"), // some stateless action which does not require logging in (or possibly even prior registration) will be granted.
         ;