+++ incubator/spring-lnurl/spring-lnurl-auth-security/src/integTest/java/org/tbk/spring/lnurl/security/test/app2/LnurlAuthTestSecurityDeprecatedConfig.java
@@ -15,8 +15,8 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.firewall.StrictHttpFirewall;
 import org.tbk.lnurl.auth.K1Manager;
-import org.tbk.lnurl.auth.LnurlAuthPairingService;
 import org.tbk.spring.lnurl.security.LnurlAuthConfigurer;
+import org.tbk.spring.lnurl.security.userdetails.LnurlAuthUserPairingService;
 
 import static org.springframework.security.config.Customizer.withDefaults;
 import static org.springframework.security.web.util.matcher.AntPathRequestMatcher.antMatcher;
@@ -28,10 +28,11 @@ import static org.springframework.security.web.util.matcher.AntPathRequestMatche
 class LnurlAuthTestSecurityDeprecatedConfig implements WebSecurityCustomizer {
 
     @NonNull
-    private final LnurlAuthPairingService pairingService;
+    private final LnurlAuthUserPairingService pairingService;
 
     @NonNull
     private final K1Manager k1Manager;
+
     @NonNull
     private final UserDetailsService userDetailsService;
 
@@ -57,7 +58,8 @@ class LnurlAuthTestSecurityDeprecatedConfig implements WebSecurityCustomizer {
                         ).permitAll()
                         .anyRequest().authenticated()
                 )
-                .with(LnurlAuthConfigurer.create(k1Manager, pairingService), it -> {});
+                .with(LnurlAuthConfigurer.create(k1Manager, pairingService), it -> {
+                });
 
         return http.build();
     }