+++ incubator/spring-lnurl/spring-lnurl-auth-security/src/integTest/java/org/tbk/spring/lnurl/security/test/shared/TestPairingService.java
@@ -5,16 +5,19 @@ import com.google.common.cache.CacheBuilder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.provisioning.UserDetailsManager;
 import org.tbk.lnurl.auth.K1;
 import org.tbk.lnurl.auth.LinkingKey;
-import org.tbk.lnurl.auth.LnurlAuthPairingService;
+import org.tbk.lnurl.auth.SignedLnurlAuth;
+import org.tbk.spring.lnurl.security.userdetails.LnurlAuthUserPairingService;
 
 import java.time.Duration;
 import java.util.Optional;
 
 @RequiredArgsConstructor
-public class TestPairingService implements LnurlAuthPairingService {
+public class TestPairingService implements LnurlAuthUserPairingService {
     private final UserDetailsManager userDetailsManager;
 
     private final Cache<K1, LinkingKey> cache = CacheBuilder.newBuilder()
@@ -22,27 +25,31 @@ public class TestPairingService implements LnurlAuthPairingService {
             .build();
 
     @Override
-    public Optional<LinkingKey> findPairedLinkingKeyByK1(K1 k1) {
-        return Optional.ofNullable(cache.getIfPresent(k1));
+    public UserDetails pairUserWithK1(SignedLnurlAuth auth) {
+        UserDetails user = createUserIfMissing(auth.getLinkingKey());
+
+        cache.put(auth.getK1(), auth.getLinkingKey());
+
+        return user;
     }
 
     @Override
-    public boolean pairK1WithLinkingKey(K1 k1, LinkingKey linkingKey) {
-        createUserIfMissing(k1, linkingKey);
-
-        cache.put(k1, linkingKey);
-
-        return false;
+    public Optional<UserDetails> findPairedUserByK1(K1 k1) {
+        return Optional.ofNullable(cache.getIfPresent(k1))
+                .map(it -> userDetailsManager.loadUserByUsername(it.toHex()));
     }
 
-    private void createUserIfMissing(K1 k1, LinkingKey linkingKey) {
-        boolean userExists = userDetailsManager.userExists(linkingKey.toHex());
-        if (!userExists) {
-            userDetailsManager.createUser(User.builder()
+    private UserDetails createUserIfMissing(LinkingKey linkingKey) {
+        try {
+            return userDetailsManager.loadUserByUsername(linkingKey.toHex());
+        } catch (UsernameNotFoundException e) {
+            UserDetails user = User.builder()
                     .username(linkingKey.toHex())
-                    .password(k1.toHex()) // password must not be null -_-
+                    .password("") // password must not be null -_-
                     .authorities(new SimpleGrantedAuthority("ROLE_LNURL_AUTH_TEST_USER"))
-                    .build());
+                    .build();
+            userDetailsManager.createUser(user);
+            return user;
         }
     }
 }